# HTTP/3: QPACK decoder performance