--- a/src/Shared/test/Shared.Tests/runtime/Http3/QPackDecoderTest.cs
+++ b/src/Shared/test/Shared.Tests/runtime/Http3/QPackDecoderTest.cs
@@ -101,6 +101,39 @@ namespace System.Net.Http.Unit.Tests.QPack
         }
 
         [Fact]
+        public void DecodesAuthority_Value()
+        {
+            byte[] encoded = Convert.FromBase64String("AADR11AOMTI3LjAuMC4xOjUwMDHBNwFhbHQtdXNlZA4xMjcuMC4wLjE6NTAwMQ==");
+
+            KeyValuePair<string, string>[] expectedValues = new[]
+            {
+                new KeyValuePair<string, string>(":method", "GET"),
+                new KeyValuePair<string, string>(":scheme", "https"),
+                new KeyValuePair<string, string>(":authority", "127.0.0.1:5001"),
+                new KeyValuePair<string, string>(":path", "/"),
+                new KeyValuePair<string, string>("alt-used", "127.0.0.1:5001"),
+            };
+
+            TestDecodeWithoutIndexing(encoded[2..], expectedValues);
+        }
+
+        [Fact]
+        public void DecodesAuthority_Empty()
+        {
+            byte[] encoded = Convert.FromBase64String("AAA3ADptZXRob2QDR0VUNTpwYXRoAS83ADpzY2hlbWUEaHR0cDcDOmF1dGhvcml0eQA=");
+
+            KeyValuePair<string, string>[] expectedValues = new[]
+            {
+                new KeyValuePair<string, string>(":method", "GET"),
+                new KeyValuePair<string, string>(":scheme", "http"),
+                new KeyValuePair<string, string>(":authority", ""),
+                new KeyValuePair<string, string>(":path", "/"),
+            };
+
+            TestDecodeWithoutIndexing(encoded[2..], expectedValues);
+        }
+
+        [Fact]
         public void DecodesLiteralFieldLineWithLiteralName_HuffmanEncodedValue()
         {
             byte[] encoded = _literalFieldLineWithLiteralName
@@ -118,6 +151,26 @@ namespace System.Net.Http.Unit.Tests.QPack
                 .ToArray();
 
             TestDecodeWithoutIndexing(encoded, _headerNameString, _headerValueString);
+        }
+
+        [Fact]
+        public void DecodesLiteralFieldLineWithLiteralName_LargeValues()
+        {
+            int length = 0;
+            Span<byte> buffer = new byte[1024 * 1024];
+            QPackEncoder.EncodeLiteralHeaderFieldWithoutNameReference(":method", new string('A', 8192 / 2), buffer.Slice(length), out int bytesWritten);
+            length += bytesWritten;
+            QPackEncoder.EncodeLiteralHeaderFieldWithoutNameReference(":path", new string('A', 8192 / 2), buffer.Slice(length), out bytesWritten);
+            length += bytesWritten;
+            QPackEncoder.EncodeLiteralHeaderFieldWithoutNameReference(":scheme", "http", buffer.Slice(length), out bytesWritten);
+            length += bytesWritten;
+
+            TestDecodeWithoutIndexing(buffer.Slice(0, length).ToArray(), new[]
+            {
+                new KeyValuePair<string, string>(":method", new string('A', 8192 / 2)),
+                new KeyValuePair<string, string>(":path", new string('A', 8192 / 2)),
+                new KeyValuePair<string, string>(":scheme", "http")
+            });
         }
 
         public static readonly TheoryData<byte[]> _incompleteHeaderBlockData = new TheoryData<byte[]>
@@ -146,11 +199,29 @@ namespace System.Net.Http.Unit.Tests.QPack
 
         private static void TestDecodeWithoutIndexing(byte[] encoded, string expectedHeaderName, string expectedHeaderValue)
         {
-            TestDecode(encoded, expectedHeaderName, expectedHeaderValue, expectDynamicTableEntry: false, byteAtATime: false);
-            TestDecode(encoded, expectedHeaderName, expectedHeaderValue, expectDynamicTableEntry: false, byteAtATime: true);
+            KeyValuePair<string, string>[] expectedValues = new[] { new KeyValuePair<string, string>(expectedHeaderName, expectedHeaderValue) };
+
+            TestDecodeWithoutIndexing(encoded, expectedValues);
         }
 
-        private static void TestDecode(byte[] encoded, string expectedHeaderName, string expectedHeaderValue, bool expectDynamicTableEntry, bool byteAtATime)
+        private static void TestDecodeWithoutIndexing(byte[] encoded, KeyValuePair<string, string>[] expectedValues)
+        {
+            TestDecode(encoded, expectedValues, expectDynamicTableEntry: false, bytesAtATime: null);
+
+            for (int i = 1; i <= 20; i++)
+            {
+                try
+                {
+                    TestDecode(encoded, expectedValues, expectDynamicTableEntry: false, bytesAtATime: i);
+                }
+                catch (Exception ex)
+                {
+                    throw new Exception($"Error when decoding with chunk size {i}.", ex);
+                }
+            }
+        }
+
+        private static void TestDecode(byte[] encoded, KeyValuePair<string, string>[] expectedValues, bool expectDynamicTableEntry, int? bytesAtATime)
         {
             var decoder = new QPackDecoder(MaxHeaderFieldSize);
             var handler = new TestHttpHeadersHandler();
@@ -158,23 +229,38 @@ namespace System.Net.Http.Unit.Tests.QPack
             // Read past header
             decoder.Decode(new byte[] { 0x00, 0x00 }, endHeaders: false, handler: handler);
 
-            if (!byteAtATime)
+            if (bytesAtATime == null)
             {
                 decoder.Decode(encoded, endHeaders: true, handler: handler);
             }
             else
             {
-                // Parse data in 1 byte chunks, separated by empty chunks
-                for (int i = 0; i < encoded.Length; i++)
+                int chunkSize = bytesAtATime.Value;
+                
+                // Parse data in chunks, separated by empty chunks
+                for (int i = 0; i < encoded.Length; i += chunkSize)
                 {
-                    bool end = i + 1 == encoded.Length;
+                    int resolvedSize = Math.Min(encoded.Length - i, chunkSize);
+                    bool end = i + resolvedSize == encoded.Length;
+
+                    Span<byte> chunk = encoded.AsSpan(i, resolvedSize);
 
                     decoder.Decode(Array.Empty<byte>(), endHeaders: false, handler: handler);
-                    decoder.Decode(new byte[] { encoded[i] }, endHeaders: end, handler: handler);
+                    decoder.Decode(chunk, endHeaders: end, handler: handler);
                 }
             }
 
-            Assert.Equal(expectedHeaderValue, handler.DecodedHeaders[expectedHeaderName]);
+            foreach (var expectedValue in expectedValues)
+            {
+                try
+                {
+                    Assert.Equal(expectedValue.Value, handler.DecodedHeaders[expectedValue.Key]);
+                }
+                catch (Exception ex)
+                {
+                    throw new InvalidOperationException($"Error when checking header '{expectedValue.Key}'.", ex);
+                }
+            }
         }
     }
 
@@ -185,6 +271,11 @@ namespace System.Net.Http.Unit.Tests.QPack
 
         void IHttpHeadersHandler.OnHeader(ReadOnlySpan<byte> name, ReadOnlySpan<byte> value)
         {
+            if (name.Length == 0)
+            {
+                throw new InvalidOperationException("Header with length zero.");
+            }
+
             string headerName = Encoding.ASCII.GetString(name);
             string headerValue = Encoding.ASCII.GetString(value);