Before:

```
|                                           Method |          Mean |        Error |       StdDev |         Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------------- |--------------:|-------------:|-------------:|-------------:|------:|------:|------:|----------:|
|          DecodeHeaderFieldLine_LargeLiteralValue |  73,798.03 ns | 1,406.262 ns | 1,504.685 ns |     13,550.5 |     - |     - |     - |         - |
| DecodeHeaderFieldLine_LargeLiteralValue_Multiple | 409,622.33 ns | 6,067.124 ns | 5,378.347 ns |      2,441.3 |     - |     - |     - |         - |
|            DecodeHeaderFieldLine_Static_Multiple |      61.78 ns |     0.743 ns |     0.620 ns | 16,187,009.9 |     - |     - |     - |         - |
```

After:

```
|                                           Method |      Mean |    Error |   StdDev |         Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------------- |----------:|---------:|---------:|-------------:|------:|------:|------:|----------:|
|          DecodeHeaderFieldLine_LargeLiteralValue |  82.09 ns | 1.526 ns | 1.191 ns | 12,182,061.0 |     - |     - |     - |         - |
| DecodeHeaderFieldLine_LargeLiteralValue_Multiple | 365.05 ns | 5.234 ns | 4.640 ns |  2,739,381.0 |     - |     - |     - |         - |
|            DecodeHeaderFieldLine_Static_Multiple |  63.02 ns | 1.269 ns | 1.187 ns | 15,867,472.9 |     - |     - |     - |         - |
```